CTPL templating engine VAPI.
These are bindings (probably incomplete) for CTPL, a C template library
http://ctpl.tuxfamily.org/

Basic functionnalities are bound nicely, but I am pretty sure that it
does not cover the whole library.

This is an orignal work from Matias De la Puente
https://mail.gnome.org/archives/vala-list/2011-December/msg00045.html
and is used by the Valum framework as its main templating engine.