Adds bindings for server.h, server_list.h in libmemcached.vapi.
Binds 'parse.h', 'server.h', 'server_list.h' and 'quit.h' to provide
more server and instance related capabilities.

Adds in_port_t definition with IntegerType annotation to make it
possible to assign an integer for the port value.

Binds MEMCACHED_DEFAULT_PORT and use it as a default value if the port
is the last argument.