Add transform () to augeas.vapi
transform() allows individual lens and configuration file combinations to be loaded.
So InitFlags.NO_MODL_AUTOLOAD can now be used when initializing the library resulting
in much quicker start up times. Reducing the time from several seconds without the
flag to a fraction of a second with the flag.

This patch also adds sensible default values for print () and srun () and adds valadoc
comments for a number of functions.