msgpack: Fixes for 'Unpacker' and 'Unpacked'
Add a constructor to 'Unpacked' to call 'msgpack_unpacked_init'.

Let 'off' be 'null' by default in 'Unpacked.next'.

Add 'buffer_consumed' to 'Unpacked' which has to be invoked once we
write to 'buffer' with the number of bytes written.

Make 'data' in 'Unpacker' return an unowned 'Object' since it is held by
the unpacker.